per-tag settings
An implementation for dvtm of the dwm pertag patch. With this patch, the
below settings are chosen and remembered individually for each tag. This
is in contrast to the default functionality where changes to these
settings are applied across all tags. Each tag starts out with the
default settings from your config.h file.
    * Layout
    * Number of clients in the master window
    * Percentage of the screen used by the master window
    * Status bar position and visibility
    * Status of the RunInAll feature
tagging paradigm workflow -
When toggling to view multiple tags, whichever tag is viewed first
determines the settings for all the pertag features. From there you
toggle additional tags on and off, and the pertag feature settings
remain unchanged. If you then switch and choose to view a different
individual tag, that new tag's settings become active for all the pertag
features. From there, you again toggle additional tags on and off. Now
you're able to toggle between the two tag-views you've set up by using
the viewprevtag action (MOD+v+TAB) and the pertag feature settings
remain as they were. In this way, you are not forced to reset all the
pertag features every time you toggle with the viewprevtag action.
workspace paradigm workflow -
Though dvtm was designed for the tagging paradigm (above), it also works
very well in a workspace paradigm. With this, you primarily use the view
and tag actions rather than the toggleview and toggletag actions. All
pertag feature settings that are set while viewing an individual tag are
remembered for that tag.

Signed-off-by: Vadym Kochan <vadim4j@gmail.com>